==> 00-acknowledgements.Rmd <==
```{r, echo = FALSE, results = "asis", include = knitr:::pandoc_to() %in% c("docx", "html")}
cat("
Acknowledgements {-}
========================================================================

")

cat(readr::read_lines("00-text-of-acknowledgements.md"))
```

\mainmatter
