==> 02-specific-aims.Rmd <==
Overview and aims
========================================================================

Individual differences in language ability are apparent as soon as
children start talking, but it is difficult to identify children at risk
for language delay or disorder. Recent work suggests word recognition
efficiency—that is, how well children map incoming speech to words—may
help identify early differences in children’s language trajectories.
Children learn spoken language by listening to caregivers, so children
who are faster at recognizing words have an advantage for word learning.
This view is borne out by some studies suggesting that children who are
faster at processing words show greater vocabulary gains months later
[e.g., @Weisleder2013].

We do not know, however, how word recognition itself develops over time
within a child. This is an important open question because word
recognition may provide a key mechanism for understanding how individual
differences emerge in word learning and persist into early language
development. Without a developmental account of word recognition, we
lack the context for understanding individual differences in lexical
processing. Thus, even the big-picture questions are unclear: Do early
differences persist over time so that faster processors remain
relatively fast later in childhood? Or, is such a question ill-posed
because the magnitude of the differences among children shrink with age?
In this dissertation, I address this gap in knowledge by analyzing three
years of word recognition data collected in a recently completed
longitudinal study of 160 children.

In particular, I examine the development of *familiar word
recognition*, *lexical competition,* and *fast referent selection* (the
ability to map novel words to novel objects in the moment). Through
these analyses, I develop a fine-grained description of how the
dynamics of word recognition change year over year, and I document how
differences in word recognition performance relate to other child-level
measures (such as vocabulary and speech perception). 

Study 1: Familiar word recognition and lexical competition {-}
------------------------------------------------------------------------

***Specific Aim\:* To characterize the development of familiar word
recognition and lexical competition, I analyze data from a Visual World
Paradigm experiment, conducted at age 3, age 4, and age 5.**

In these eyetracking experiments, children were presented with four
images of familiar objects and heard a prompt to view one of the images.
The four images included a target word (e.g., *bell*), a semantically
related word (*drum*), a phonologically similar word (*bee*), and an
unrelated word (*swing*). In [Chapter \@ref(fam-rec)](#fam-rec), I use a
series of growth curve analyses to describe how children’s familiar word
recognition develops year over year. Children in this cohort cover a
range of vocabulary scores at age 3, and this variability allows me to
investigate individual differences in vocabulary and word recognition
over time and assess the predictive value of these measures. Of interest
was how individual differences at age 3 persisted into age 5 and how these
differences related to vocabulary measures at later ages. In
[Chapter \@ref(lex-competitors)](#lex-competitors), I examine the
children’s looks to the distractors to study the developmental course of
lexical competition from similar sounding and similar meaning words.
Increases in sensitivity to competing words reveal how lexical
competition effects emerge as a byproduct of learning new words and
developing more efficient phonological and lexical representations. As I
argue in [Chapter \@ref(aim1-discussion)](#aim1-discussion), increased
sensitivity to lexical competitors supports familiar word recognition
because children become more efficient at activating a named word *and
related words*. When children err, they become more likely to err on a
lexically relevant alternative.


Study 2: Referent selection and mispronunciations {-}
-----------------------------------------------------------------------

***Specific Aim\:* To characterize how fast referent selection develops
longitudinally, I analyze data from a looking-while-listening
mispronunciation experiment, conducted at age 3, age 4, and age 5.**

Not every word children hear are familiar to them. They may hear
entirely new words, or they may hear variations and corruptions of
familiar words. How children respond to both kinds of words is
informative, as I review in
[Chapter \@ref(aim2-introduction)](#aim2-introduction). I describe this
eyetracking experiment in detail---based on @WhiteMorgan2008 and
@MPPaper---in [Chapter \@ref(aim2-method)](#aim2-method). Children saw an
image of a familiar object and an unfamiliar object, and they heard
either a correct production of the familiar object (e.g., *soup*), a
one-feature mispronunciation of the familiar object (*shoup*), or a
novel word unrelated to either image (*cheem*). The correct productions
tested familiar word recognition and the nonwords tested fast referent
selection. The mispronunciations tested the child’s phonological
categories by showing whether the child permitted, rejected, or
equivocated about mispronunciations. 

I use growth curve analyses to study how children’s responses to the
three word types changed over time. In
[Chapter \@ref(real-nonword-selection)](#real-nonword-selection), I
examine familiar word recognition and fast referent selection for novel
words to determine which feature of lexical processing better predicts
vocabulary growth. I compare these two conditions directly to look for
dissociations or asymmetries in these forms of processing within
children as a way to empirically assess the claim that "novel word
processing (referent selection) is not distinct from familiar word
recognition" [@McMurray2012]. In
[Chapter \@ref(sensitivity-to-mispronunciations)](#sensitivity-to-mispronunciations),
I examine how children interpreted mispronunciations of the familiar
words at each age and study how individual differences in vocabulary and
speech perception related to children's responses to the
mispronunciations. I also report how children at age 5 are better able
to retain nonwords than mispronunciations of familiar words.




Summary {-}
------------------------------------------------------------------------

This project investigates how word recognition develops during the
preschool years. There has been no published research studying word
recognition longitudinally after age two. Furthermore, this project also
examines word recognition in two experimental tasks that tap into
different aspects of word recognition. Specifically, a four-image
experiment with semantic and phonological foils allows me to study how
lexical competition develops, and a two-image experiment with nonwords
and mispronunciations enables me to study how children's responses to
unfamiliar words develop over time as well.
[Chapter \@ref(both-studies-discussion)](#both-studies-discussion)
reviews the results from both studies in terms of lexical processing as
well as the main contributions of this project. Findings show how
individual differences in lexical processing change over time and reveal
how low-level mechanisms underlying word recognition mature
longitudinally in children. These findings have translational value by
studying processing abilities that subserve word learning and by
assessing the predictive relationships between early word recognition
ability and later language outcomes.
