==> 03-significance.Rmd <==
Significance
========================================================================


Public Health Significance
------------------------------------------------------------------------

Vocabulary size in preschool is a robust predictor of later language
development, and early language skills predict early literacy skills at
school entry [@Morgan2015]. By studying the mechanisms that shape word
learning, we can understand how individual differences in language
ability arise and identify strategies for closing language gaps between
children. Word recognition—the process of mapping incoming speech sounds
to known or novel words—has been shown to predict later language
outcomes. We do not know how this ability develops over time, and we do
not know when word recognition is most predictive of future outcomes.
This project will provide an integrated account of how word recognition
and its relationship with vocabulary size change from age 3 to age 5.


Scientific Significance
------------------------------------------------------------------------

### Lexical Processing Dynamics

Mature listeners recognize words by continuously evaluating incoming
speech input for possible word matches through lexical competition. The
first part of a word activates multiple candidate words in parallel, and
these candidates compete so that the best-fitting word is recognized.
For example, the onset "bee" might activate the candidates *bee*,
*beam*, *beetle*, *beak*, *beaker*, *beginning*, and so on, but an
additional "m" would narrow the candidates to just *beam*. Semantic
relationships also influence lexical processing, and cascading
phonological-semantic effects—e.g., where *castle* activates the
phonologically similar *candy* which in turn activates the semantically
related *sweet*—have been demonstrated [@Marslen-Wilson1989]. Both
low-level phonetic cues and high-level grammatical, semantic and
pragmatic information can influence this process, but the *continuous
processing of multiple competing candidates* is the essential dynamic
underlying word recognition [@Magnuson2013].

What about young children who know considerably fewer words? Eyetracking
studies with toddlers have suggested a developmental continuity between
toddlers and adult listeners. Children recognize words incrementally
[@Swingley1999], match truncated words to their intended referents
[@Fernald2001], and use information from neighboring words in a sentence
to facilitate word recognition. This information can be grammatical:
@Lew-Williams2007 found that Spanish-acquiring preschoolers can use
grammatical gender on determiners (*el* or *la*) to anticipate the word
named in a two-object word recognition task. The information can also be
subcategorical phonetic variation: We found that English-acquiring
toddlers look earlier to a named image when the coarticulatory formant
cues on word *the* predict the noun of the sentence, compared to tokens
with neutral coarticulation [@Mahr2015].

There is some evidence for lexical competition where children are
sensitive to phonological and semantic similarities among words.
@EllisWeismer2016 showed that toddlers (14–29 months old) looked less
reliably to a named image when the onscreen competitor was a
semantically related word or perceptually similar image. In @RWLPaper,
preschoolers (28–60 months old) demonstrated sensitivity to semantic and
phonological competitors in a four-image eyetracking task. @Huang2011
presented evidence of cascading semantic-phonological activation in
five-year-olds such that for a target word like *log*, children looked
more to an indirect phonological competitor like *key* (competing
through its activation of *lock*) than they looked to an unrelated image
like *carrot*. In contrast to these studies which all demonstrate
interference from similar words, @Mani2010 demonstrated cross-modal
phonological priming effects in 18-month-olds. In this study, a picture
of prime word (e.g., cat or teeth) was presented in silence; then two
images (e.g., cup and shoe) were presented, one of which was named
(*cup*). Children on average looked more to the target word (like *cup*)
when it was primed by an image of a phonological neighbor (like *cat*),
and the children performed at chance when the prime was not related to
the named word. @Mani2012 found a similar result for cascading
phonological-semantic priming with 24-month-olds: Children looked more
to a target *shoe* compared to a distractor *door* when primed by an
image of *clock*, assumed to activate *sock* which primed *shoe*.

The above studies involved young children of different ages tested under
different procedures, sometimes in different dialects and languages.
Averaging these results together, so to speak, the studies suggest that
early word recognition demonstrates some hallmarks of adult behavior:
Continuous processing of words, integration of information from
different levels of representation, and the influence of similar,
unspoken words on recognition of a word. Nevertheless, we only have a
fragmented view of how familiar word recognition and lexical competition
develop within children.

One open question is how lexical competition develops within children.
For example, do phonological similar words exert more interference
during word recognition as children grow older? As a guiding hypothesis,
we can think of word learning as a gradual process where familiarity
with a word moves from shallow receptive knowledge to deeper expressive
knowledge. In adult listeners, words compete and inhibit one another, so
that a word is truly “learned” and integrated into the lexicon when it
can influence the processing of other words [a line of reasoning
reviewed by @Kapnoula2015]. Increasing sensitivity to similar sounding
words over time would reveal that children improve their ability to
consider multiple candidates in parallel. By studying how sensitivity to
similar-sounding and similar-meaning words develop over time and within
ever-growing vocabularies, this project can reveal how children come to
process words efficiently.



### Individual Differences in Word Recognition

We have a rough understanding of the development of word recognition,
and these gaps in knowledge matter because young children differ in
their word recognition abilities. These differences are usually measured
using *accuracy* (a probability of recognizing to a word) or
*efficiency* (a reaction time or some measure of how quickly accuracy
changes over time). These differences are consequential too, as word
recognition differences correlate with other language measures
concurrently, retrospectively, and prospectively.

The best predictor of lexical processing efficiency is concurrent
vocabulary size: Children who know more words look more quickly and
reliably to a named word [e.g., @MPPaper].

This fact deserves a brief reflection: Suppose the information
processing mechanism behind word recognition were just a naïve table
search. Then this finding is somewhat puzzling: Children with larger
lexicons have to find a needle in a larger haystack—yet this apparent
liability is an advantage. That is why the search analogy is naïve. One
explanation follows from the earlier described idea about graded word
learning: Children become better at recognizing words as they learn more
words because they extract regularities and discover similarities among
words and develop more efficient lexical representations—the haystack
develops regularity and becomes easier to search.

Although it is a robust predictor of word recognition, vocabulary size
is nonspecific. For lexical processing dynamics, vocabulary size can be
considered an indicator for the organization and efficiency of a child’s
lexicon, but it also correlates with other (meaningful) differences.
Vocabulary is related to differences in speech perception
[@Cristia2014_Review] and environmental factors like language input
[e.g., @HartRisley; @Hoff2003]. For instance, measures of speech
perception at 6–8 months predict vocabulary size at 24 months [e.g.,
@Tsao2004; @Kuhl2008], so processing predicts future vocabulary predicts
concurrent processing.

A related complication is the apparent predictive validity of word
recognition measures. @MarchmanFernald2008 found that vocabulary size
and lexical processing efficiency at age 2 jointly predicted working
memory scores and expressive language scores at age 8. This result would
suggest domain-general processing advantages influence word learning.
@Fernald2012 also found that late talkers who looked more quickly to a
named word at 18 months showed larger gains in vocabulary by 30 months
compared to late-talkers who looked more slowly at 18 months.
@Weisleder2013 found that lexical processing and language input at
19 months predict vocabulary size at 25 months and that lexical
processing mediated the effect of language input.

Word recognition efficiency and vocabulary size are interconnected
measures with concurrent and predictive associations. This project can
clarify this relationship by examining the co-development of word
recognition, vocabulary size, and speech perception. In particular, I
will ask how individual differences in word recognition change over time
alongside differences in vocabulary. I can also which features of word
recognition (fast referent selection, lexical competition, etc.) are
most predictive of vocabulary outcomes at age 5. The additional measures
of speech perception can also help clarify the specific effects of
vocabulary size on word recognition.

### Summary

This project studies word recognition in children over three years, so
it will provide the first longitudinal study of word recognition in
preschoolers. Children in this cohort cover a range of vocabulary scores
at Time 1, and this variability allows one to investigate individual
differences in vocabulary and word recognition over time and assess the
predictive value of these measures. Furthermore, this project studies
word recognition in two experimental tasks that can tap into different
aspects of word recognition. Specifically, a four-image experiment with
semantic and phonological foils allows me to study how lexical
competition develops, and a two-image experiment with nonwords and
mispronunciations enables me to study how fast referent selection
develops over time as well. I will use mixed effects modeling to study
not just gross measures of accuracy or interference from distractors,
but the time course and lexical dynamics of word recognition using
growth curve analyses.
