==> 04-hypotheses.Rmd <==
Research hypotheses
========================================================================

In this section, I outline the main hypotheses I plan to examine for each
study. This section is intended to preregister the main analyses
for this project.

Study 1: Familiar word recognition and lexical competition {-}
------------------------------------------------------------------------

  - Children’s accuracy and efficiency of recognizing words will improve
    each year.

  - There are stable individual differences in lexical processing of
    familiar words such that children who are relatively fast at age 3
    remain relatively fast at age 4 and age 5.

  - However, the magnitude of these individual differences diminishes
    over time, as children converge on a mature level of performance for
    this paradigm.

  - Consequently, individual differences in word recognition at
    age 3, for example, will be more discriminating and predictive of
    age-5 language outcomes than differences at age 4 or age 5.

  - Children will become more sensitive to lexical competitors as they
    age, based on the hypothesis that children discover similarities
    among words as a consequence of learning more and more words.

  - Children will differ in their sensitivity to lexical competitors,
    and these individual differences will correlate with other
    child-level measures.



Study 2: Referent selection and mispronunciations {-}
------------------------------------------------------------------------

  - Children’s accuracy and efficiency of recognizing real words and
    fast-associating nonwords will improve each year.

  - Performance in real word recognition and fast association of
    nonwords will be highly correlated, based on the hypothesis that the
    same process (referent selection) operates in both situations.

  - Under the alternative hypothesis, real word recognition and fast
    referent selection reflect different skills with different
    developmental trajectories. Thus, if there is any dissociation
    between recognition of real words and nonwords, it will be observed
    in younger children.

  - Although these two measures will be correlated, I predict
    performance in the nonword condition will be a better predictor of
    future vocabulary growth than performance in the real word
    condition. This hypothesis is based on the idea that fast referent
    selection is a more relevant skill for learning new words than
    recognition of known words.

  - For the mispronunciations, I predict children with larger
    vocabularies (that is, older children) will be more likely to
    tolerate a mispronunciation as a production of familiar word
    compared to children with smaller vocabularies.

  - Mispronunciations that feature later-mastered sounds
    (e.g., *rice-wice*) will be more likely to be associated to novel
    objects than earlier-mastered sounds (*duck-guck*).
