==> 11-aim1-introduction.Rmd <==
Familiar word recognition {#aim1-introduction}
=======================================================================

Lexical processing dynamics
------------------------------------------------------------------------

Mature listeners recognize spoken words by continuously evaluating
incoming speech for possible word matches. The first part of a word
activates multiple candidate words in parallel. These candidates compete
as more of the speech signal enters the system, and the best-fitting
word is the favored interpretation. For example, the onset "bee" might
activate phonologically compatible candidates like *bee*, *beam*,
*beetle*, *beak*, *beaker*, *beginning*, and so on, but an additional
"m" would narrow the candidates to just *beam*. Semantic relationships
also influence lexical processing, and cascading phonological-semantic
effects—for instance, where *castle* activates the phonologically
similar *candy* which in turn activates the semantically related
*sweet*—have been demonstrated [@Marslen-Wilson1989]. Both low-level
phonetic cues and high-level grammatical, semantic and pragmatic
information can influence this process, but this *continuous processing
of multiple competing candidates* is the essential dynamic underlying
word recognition in adults [@Magnuson2013].

What about young children who know considerably fewer words? Eyetracking
studies with toddlers have suggested a developmental continuity between
toddlers and adult listeners. Children recognize words incrementally
[@Swingley1999], match truncated words to their intended referents
[@Fernald2001], and use information from neighboring words in a
sentence to facilitate word recognition. This information can be
high-level grammatical or semantic cues. @Lew-Williams2007 found that
Spanish-acquiring preschoolers can use grammatical gender on determiners
(*el* or *la*) to anticipate the word named in a two-object word
recognition task. @Borovsky2012 showed that children can use semantic
information from an agent and a verb (e.g., *the dog chased*) to
anticipate a plausible noun (*the cat*). The information can also be
low-level phonetic variation: We found that toddlers look earlier to a
named image when the coarticulatory formant cues on word *the* predicted
the noun of the sentence, compared to tokens with neutral coarticulation
[@Mahr2015].

There is some evidence for lexical competition where children are
sensitive to phonological and semantic similarities among words.
@EllisWeismer2016 showed that toddlers (14–29 months old) look less
reliably to a named image when the onscreen competitor was a
semantically related word or perceptually similar image. @Huang2011
presented evidence of cascading semantic-phonological activation in
five-year-olds such that for a target word like *log*, the children
looked more to an indirect phonological competitor like *key* (competing
through its activation of *lock*) than they looked to an unrelated image
like *carrot*.

Priming studies also reveal that children are sensitive to phonological 
similarities among words. @Mani2010 demonstrated
cross-modal phonological priming effects in 18-month-olds. In this
study, a picture of prime word (e.g., cat or teeth) was presented in
silence; then two images (cup and shoe) were presented, one of which was
named (*cup*). Children on average looked more to the target word
(*cup*) when it was primed by an image of a phonological neighbor
(*cat*), and the children performed at chance when the prime was not
related to the named word. @Mani2012 found a similar result for
cascading phonological-semantic priming with 24-month-olds: Children
looked more to a target (e.g., *shoe*) compared to a distractor (*door*)
when primed by an image (*clock*, assumed to activate *sock* which
primed *shoe*).[^semantic-priming]

@AltvaterMackensen2013 demonstrated phonological-semantic priming even
when the prime is a mispronunciation. German-learning two-year-olds
heard a prime word, and 200 ms later two images appeared onscreen (a cow,
*Kuh* and a fork, *Gabel*), one of which was labeled ("Kuh"). The prime word
was a semantically related word ("Schaf", sheep), an
onset-mispronunciation of the related word ("Faf" or "Taf"), or an
unrelated prime ("Buch", book). Children looked to the target about
equally well in the two prime conditions: approximately .62 proportion
looks to target with a normal prime versus approximately .60 for a
mispronounced prime. In contrast, they looked less to the target in the
unrelated prime condition (approximately .55). Thus, the children in
this study showed cascading activation where the mispronunciation
activates the mispronounced word which in term activates a semantically
related word.

@Chow2017 performed a very similar study to the one I present in the next chapters.
They used the Visual World Paradigm with English-learning 24-
and 30-month-olds. Children saw a 2 × 2 grid of images which included a
phonological (cohort) competitor and a semantic competitor, and they
heard a prompt to view one of the images (e.g., *Look at the bee*). On *filler*
trials, the target word and an unrelated image appear onscreen alongside the
competitors. On *test* trials, the display had two unrelated images
(*sandwich*, *dress*), a phonological competitor (*bus*), and a semantic
competitor (*cat*) and children were prompted to look at an offscreen,
unpictured target (*Look at the bee*). They found a
temporary early advantage for the
phonological competitor, so that the probability of looking to the
phonological competitor was greater than the other competitors. This
early advantage was followed by a late, more stable advantage for the
semantic competitor. Moreover, they found that increased receptive
vocabulary predicted more looks to the phonological competitor and fewer
looks to the semantic competitor. (The looks to the semantic competitor
were decreased because of the early advantage of the phonological
competitor.) Their results support a kind of cascading activation in
which phonological information comes online before semantic information.


[^semantic-priming]: @Arias-Trejo2009 is commonly cited as evidence of
semantic priming effects. Toddlers heard sentences like "I saw a cat...
dog". During the word *dog*, two images (dog and door) are presented.
The idea is that *cat* should prime looks to its semantic neighbor
*dog*. The unnatural stimulus order (a sentence followed by an isolated
single word) and a condition effect where 18-month-olds
outperformed 21-month-olds make me skeptical that semantic priming is
the most plausible explanation of those results.

The above studies involved young children of different ages tested under
different procedures, sometimes in different dialects and languages.
Averaging these results together, so to speak, the studies suggest that
early word recognition demonstrates some hallmarks of adult behavior:
Continuous processing of words, integration of information from
different levels of representation, and the influence of similar,
unspoken words on the recognition of a word. Nevertheless, we only have a
fragmented view of how familiar word recognition develops within children.

One open question is how lexical competition develops in young
listeners. For example, how and when do phonologically or semantically
similar words exert their influence on word recognition? @Chow2017
provide a promising first step, in which two-year-olds looked to the
phonological and semantic relatives of a named word. (Yet I am skeptical
of any word recognition study where a target word is absent and absent
for many trials.) As a guiding
hypothesis, we can think of word learning as a gradual process where
familiarity with a word moves from shallow receptive knowledge to deeper
expressive knowledge. In adult listeners, words compete and they inhibit
one another, so that a word is truly “learned” (integrated into the
lexicon) when it can influence the processing of other words [a line of
reasoning reviewed by @Kapnoula2015]. Increasing sensitivity to similar
sounding or similar meaning words over time would reveal that children
more thoroughly learn familiar words with age. 




Individual differences in word recognition
------------------------------------------------------------------------

We have a rough understanding of the development of word recognition,
and these gaps in knowledge matter because young children differ in
their word recognition abilities. These differences are usually measured
using *accuracy* (a probability of recognizing a word) or
*efficiency* (a reaction time or some measure of how quickly accuracy
changes over time). These differences are consequential too, as word
recognition differences correlate with other language measures
concurrently and prospectively.

Many studies highlight the predictive power of word recognition ability.
@MarchmanFernald2008 found that vocabulary size and lexical processing
efficiency at age 2 jointly predicted working memory scores and
expressive language scores at age 8. @Fernald2012 found that late
talkers who looked more quickly to a named word at 18 months showed
larger gains in vocabulary by 30 months compared to late talkers who
looked more slowly at 18 months. @Weisleder2013 found that lexical
processing and language input at 19 months predicted vocabulary size
at 25 months and that lexical processing mediated the effect of
language input---the basic idea being that rich language input builds up
word recognition ability which in turn supports word learning. @Lany2017
found a direct link between lexical processing and word
learning: 18-month-olds and 30-month-olds who were faster at recognizing
familiar words were also more accurate at recognizing novel words in a
word-learning task. Thus, children who are better at recognizing words
learn more words over time and perform better at word-learning tasks.

Word recognition performance predicts future language outcomes, so we
conclude that individual differences in word recognition are important.
But we do not know how word recognition develops within children, so we
have no context for evaluating these individual differences. Are these
differences in lexical processing persistent over development? Is word
recognition a skill where most children catch up and converge on a
mature range of performance by a certain age? 


The current study
------------------------------------------------------------------------

In the previous two sections, I outlined two gaps in knowledge. The
first is that we do not have a clear understanding of how the mechanisms
underlying word recognition change in early childhood. We know that children
show plenty of adult-like features of word recognition, but each of
these findings is an isolated fact. What we need is a coherent set of
facts that shows how specific features of word recognition change with age.
The second gap is that although we know that individual differences in
word recognition are predictive of later outcomes, we do not have a
developmental picture of these individual differences.

In this study, I tackle these two lines of research: The development of
lexical competition effects and individual differences in familiar word
recognition. I report the results of a longitudinal study of word
recognition in preschoolers at age 3, age 4, and age 5. The study is
described in detail in [Chapter \@ref(aim1-method)](#aim1-method).
Briefly stated, this experiment tested word recognition by presenting
prompts like "find the horse" and recording children's looks to an array
of four images. The array of images included the target, a phonological
competitor, a semantic competitor, and an unrelated image. In 
[Chapter \@ref(fam-rec)](#fam-rec), I analyze the development patterns of
familiar word recognition (looks to the target) and how individual
differences change over time. I hypothesized that children would show
stable individual differences over time, but the range and magnitude of
these differences would get smaller as children grew older. I also
examine which word recognition measures correlate with future vocabulary
size to test how word recognition behavior predicts later outcomes.
In [Chapter \@ref(lex-competitors)](#lex-competitors), I study how the
phonological and semantic competitors influence word recognition, and I
test the prediction that children will become more sensitive to
competitors as they grow older. Finally, in
[Chapter \@ref(aim1-discussion)](#aim1-discussion), I link these two
lines of research together and describe both sets of results in terms of
lexical processing dynamics, and [Chapter \@ref(aim1-h-check)](#aim1-h-check)
reviews the results of my pre-analysis research hypotheses.



<!-- One first step towards understanding differences in word recognition is -->
<!-- the following observation: The best predictor of lexical processing -->
<!-- efficiency in children is concurrent vocabulary size. That is, children -->
<!-- who know more words look more quickly and reliably to a named word -->
<!-- [e.g., @MPPaper]. This fact deserves a brief reflection: Suppose the -->
<!-- information-processing mechanism behind word recognition were just a -->
<!-- naïve table search. Then this finding is somewhat puzzling: Children -->
<!-- with larger lexicons have to find a needle in a larger haystack—yet this -->
<!-- apparent liability is an advantage. That is why the search analogy is -->
<!-- naïve. One explanation follows from the earlier described idea about -->
<!-- graded word learning: Children become better at recognizing words as -->
<!-- they learn more words because they extract regularities and discover -->
<!-- similarities among words and develop more efficient lexical -->
<!-- representations—the haystack develops regularity and becomes easier to -->
<!-- search. The development of word recognition requires learning similarities among words and therefore requires -->




<!-- {#aim-discussion} -->

<!-- It is not clear what mechanisms explain the relationship between word -->
<!-- recognition and later outcomes. Children who are faster at recognizing -->
<!-- words could be better word learners:  -->







<!-- Taken together, these results are inconclusive. -->

<!-- This result would -->
<!-- suggest word recognition reflects domain-general processing abilities that influence word learning. -->


<!-- In [MahrEdwards2018], we evaluated -->
<!-- language input and lexical processing too. We found that language input -->
<!-- and word recognition at age 3 efficiency predicted growth in age-4 -->
<!-- receptive vocabulary, but only word recognition predicted growth in -->
<!-- age-4 expressive vocabulary. We concluded that receptive vocabulary -->


<!-- Children differ in word-recognition ability in ways that  -->


<!-- Although it is a robust predictor of word recognition, vocabulary size -->
<!-- is nonspecific. For lexical processing dynamics, vocabulary size can be -->
<!-- considered an indicator for the organization and efficiency of a child’s -->
<!-- lexicon, but it also correlates with other (meaningful) differences. -->
<!-- Vocabulary is related to differences in speech perception -->
<!-- [@Cristia2014_Review] and environmental factors like language input -->
<!-- [e.g., @HartRisley; @Hoff2003]. For instance, measures of speech -->
<!-- perception at 6–8 months predict vocabulary size at 24 months [e.g., -->
<!-- @Tsao2004; @Kuhl2008], so processing predicts future vocabulary predicts -->
<!-- concurrent processing. -->



<!-- Word recognition efficiency and vocabulary size are interconnected -->
<!-- measures with concurrent and predictive associations. This project can -->
<!-- clarify this relationship by examining the co-development of word -->
<!-- recognition, vocabulary size, and speech perception. In particular, I -->
<!-- will ask how individual differences in word recognition change over time.  -->

<!-- By studying how sensitivity to -->
<!-- similar-sounding and similar-meaning words develop over time and within -->
<!-- ever-growing vocabularies, this project can reveal how children come to -->
<!-- process words efficiently. -->
