==> 18-aim1-general-discussion.Rmd <==
General discussion {#aim1-discussion}
=======================================================================

This study examined the development of familiar word recognition over
the preschool years. The word recognition data came from a visual-world
eyetracking experiment which recorded children's fixations to images in
response to prompts like *see the bear*. The trials featured a target
noun (e.g., *bear*) along with a phonological competitor (*bell*), a
semantic competitor (*horse*), and an unrelated image (*ring*). To
describe children's word recognition ability, I analyzed how the
probability of fixating on the target image changed over the time course
of a trial. The presence of the competitor images also allowed
additional analyses about children's sensitivities to the phonological
and semantic competitors. The experiment was conducted as part of a
three-year longitudinal study; children were 28–39 months-old at the
age 3 visit, 39–52 at age 4, and 51–65 at age 5. The longitudinal
design allowed me to describe developmental changes in word recognition.



How to improve word recognition
-----------------------------------------------------------------------

Children showed year-over-year improvements in word recognition, as
measured by average looking probabilities, peak looking probabilities,
and the rate of change in looking probabilities. Children became more
reliable, less uncertain, and faster at recognizing familiar words as
they grew older. At the same time, children also became more sensitive
to the phonological and semantic competitors, compared to the unrelated
images. With each year, children looked more to the target image, but
when they erred, they were more likely to err on a lexically relevant
word.

We can interpret these developmental patterns in terms of lexical
activation and processing dynamics. In this task, children hear a stream
of speech and activate some phonetic, phonological, lexical, and
semantic representations that match the speech input. As they hear more
of a word, the activation builds until a particular word is favored, and
children shift their gaze onto the named image. Let's imagine that we
have to engineer this system. To make word recognition more efficient,
we have to find ways to increase the relative activation of the correct
word. In particular, we can boost the strength of connections so that
activation can propagate more quickly through the system, and we can
also allow inhibition among competing words so that the correct word can
win out over its competitors more quickly.

The results from these studies indicate that children become more
efficient at activating the target word *and related words* over the
preschool years. As they grew older, children were faster to look at a
named image and more likely to fixate on the phonological competitor
(compared to the unrelated image). These two findings reflect changes in
how partial acoustic information can propagate to activate
phonologically plausible words. The phonological competitors shared the
same syllable onset as the target noun (e.g., *dress*–*drum*), so the
early part of the word matched both words. That children became
more sensitive to the phonological competitor means that they learned
and somehow encoded the phonological similarities among words because
part of a word could activate a neighborhood of phonologically plausible
matches. This developmental change supports faster word recognition
because the listener can channel activation to relevant words more
quickly. A similar line of reasoning applies to the semantic
competitors: Relative looks to the semantic competitors increased with
age, suggesting that children had learned semantic connections among
words and activated semantically related words during word recognition.

The other mechanism we might tune to improve word recognition is
inhibition. Children's looks to the phonological or semantic competitors
were temporary: Looks increase to some peak level and then quickly
decrease. Behaviorally, the drop in looking probability reflects the
rejection of an interpretation: for example, a child hears "dr", shifts
looks to *dress*, but hears "um", revises the interpretation and jumps
to *drum*. We can read these corrections as evidence for an
inhibitory process: Corrections indicate a change in relative activation
where a different word overrides an initial interpretation. But the
evidence for *developmental changes* in lexical inhibition from these
data was scant. The rate of rejection of the phonological
competitor---that is, how quickly looks fall from their peak value---did
not change from age 4 to age 5, although the rate did increase for the
semantic competitor from age 4 to age 5. Preschoolers did demonstrate
inhibition by revising their interpretations of nouns, but there were no
clear developmental changes in inhibition.

Previous simulation work can help identify more specific mechanisms at
play. @McMurray2010 used the TRACE model of word recognition [@TRACE]
to simulate looks to a target and phonological competitors (cohorts and
rimes) in adolescents with specific language impairment. The authors
tuned a number of model parameters and analyzed how those changes
affected simulated looks to the target and competitors. In the current
dataset, I observed a developmental trend where the relative looks to
the phonological competitors peak higher each year. In those TRACE
simulations, looks to the cohort competitor peak higher if 1) the rate
of lexical activation increased, 2) the rate of lexical decay decreased,
or 3) strength of lexical inhibition decreased. Of these options, the
growth curve for the decrease in lexical inhibition best matches the
shape of the current data. The similarity does not mean that children
inhibited words any less as they grew older. That would be too
simplistic: Developmental changes in preschoolers are the result of
simultaneous changes in many mechanisms. But those simulation results
suggest that an *increase* in lexical inhibition is *not* one of the
key developmental changes in preschoolers' word recognition.



Learn words and learn connections between words
-----------------------------------------------------------------------

Preschoolers showed increased activation of the target noun
and semantically and phonologically related words but little
developmental change in lexical inhibition. Paired with the findings
from older children, these results lead to a compelling developmental
story. @Rigler2015 compared 9- and 16-year-olds on a Visual World word
recognition experiment with phonological (cohort and rime) competitors.
The younger children were slower to look to the target image and showed
more looks to the competitors. The implications are that children's word
recognition is still developing in late childhood and that in
particular, children's inhibition of lexical competitors became stronger
with age.

The current study with 3-, 4-, and 5-year-olds followed a different
pattern: Relative looks to the competitor images increased with age.
Taken together, these two studies suggest an interesting progression for the
development of lexical processing. During the preschool years, children
learn many, many words, and they establish phonological and semantic
connections between these words. These connections support the immediate
activation of neighborhoods of related words. Later childhood, based on the
@Rigler2015 findings, then is a time for refinement of those
connections so that sensitivity to the competitors decreases. This
refinement could follow from more selective activation channels,
increased lexical inhibition, changes in resting activation (to favor
more frequent words), or likely a combination of these factors.



Individual differences are most important at younger ages
-----------------------------------------------------------------------

Another dimension of this study concerned individual differences in word
recognition. Some children were faster or more accurate during word
recognition, and these children also were more likely to be faster or more
accurate at later ages. The magnitude of these differences diminished
over time, as children approached a more mature level of
performance. 

In terms of lexical processing dynamics, we might think of early
differences as reflecting early differences in the burgeoning lexicon.
Children may have different numbers of words, different degrees of
experience with some words, less established connections among words,
and at a lower level, different phonetic and speech perception
abilities, given the links between speech perception in infancy and
early vocabulary development [@Cristia2014_Review]. Differences in
word recognition are greatest early on in development because this is
when the differences among children's lexicons are greatest. The task of
learning new words, and more importantly, of developing representations
and associations to organize words normalizes the early differences
among children's lexicons. That pressure would make the
overall variability among children decrease over time while still
preserving a relative ordering among children.


We can also interpret the predictive power of word recognition measures
in terms of lexical processing and lexical organization. Correlations
between word recognition performance and future vocabulary were
strongest for the age-3 growth curve features, particularly for the
peak probability of looking to the target word. The peak probability
measures the overall certainty in word recognition and how strongly the
target word is activated. Children with more efficient representations
of familiar words at age 3 have a stronger foundation for encoding and
integrating future words, and as a result, they showed larger
vocabularies at age 4 and age 5. 

Initially, I had expected processing *speed*---as approximated by growth
curve slopes---to be the most predictive measure of vocabulary growth.
Children who can more quickly recognize words, the reasoning goes, can
take in information more quickly and devote extra processing resources
towards learning.[^fernald] Processing speed was indeed correlated
with future vocabulary size, yet peak probability was a stronger
predictor of future vocabulary size. Granted, these two processing
measures are highly related; to hit a higher peak by time *x*, a growth
curve needs to start from higher baseline or have steeper slope. The
idea of uncertainty suggests an alternative explanation of the predictive
power of word recognition: Children who are more accurate (or less
uncertain) during word recognition can extract and activate *more
information* from the speech signal.

[^fernald]: "The infant who identifies familiar words more quickly has
more resources available for attending to subsequent words, with
advantages for learning new words later in the sentence, as well as for
tracking distributional information about relations among words... Being
slow to identify the referent of a familiar word could interfere with
lexical activation and impede success in tracking distributional
regularities and managing attentional resources in real time (Evans,
Saffran, & Robe-Torres, 2009)" [@Fernald2012, p. 217].


Limitations and implications
------------------------------------------------------------------------

The discussion of processing speed and word recognition certainty
highlights one limitation of this research: The experiment's four-image,
eyetracking-based design meant that a clean measure of processing speed
was not feasible. Other eyetracking studies with two images can use the
latency of how long it takes the child to shift between images as a
measure like reaction time. This approach does not translate to the
four-image design, as children can visit multiple images on their way to
the target. Visual World studies with older participants can obtain an
explicit reaction time measure by means of a mouse click or tap on a
touchscreen, but those additional task demands may not translate to
young children like those in this study. Thus, this study could not
address directly whether the predictive power of word recognition
performance reflects a more developed lexicon, a general
reaction-time-like speed advantage, or both.

The lack of an explicit selection behavior, such as a mouse click, also
means that word recognition accuracy was never directly measured but
rather inferred. As a result, the interpretation of peak looking
probability as a measure of word recognition certainty comes with a
caveat: It reflects certainty averaging over many familiar words *and
maybe a few unfamiliar words*. The idea is as follows: Suppose at age 3 a
child does not know four of words well. If they had to click or tap
an image, they would have to guess on these trials. We could exclude
those trials where they guessed incorrectly, leaving just the trials
where the child correctly recognized the word. In this scenario, we
would be more justified in interpreting a growth curve peak as a measure
of certainty during familiar word recognition because trials involving
incorrectly identified words had been excluded. (It bears mentioning
that explicit selection behaviors during the experiment are just one way
to test a child's knowledge of items; another is a receptive
vocabulary test after the experiment which checks whether the child can
point to the words from the experiment.)

As it stands here, there is no clear way to tease apart whether the
lower growth curve peaks at age 3 reflected greater uncertainty during
lexical processing or a greater number of words being unfamiliar (or
unknown) to the children. I favor the former interpretation because
these were highly familiar words and because children's word recognition
improved from age 4 to age 5. We piloted the images/words in two
preschool classrooms, using only items that were at least 80%
recognizable to children. But even if some words were unfamiliar at
age 3, the number of unfamiliar words at age 4 was likely to be very
small and therefore unknown words would have exerted a minimal effect on
the lexical processing measures. The average peak looking probabilities
increased by about .13 at age 4 (from .55 to .68) and by about .09 at
age 5 (to .77)---the magnitude of these changes are comparable. Because
children also showed improvements at age 5, when the effect of unknown
words would be very small, age-related improvements in word recognition
certainty likely reflect changes in lexical processing, as opposed to 
changes in the average mixture of known and unknown words.

The experimental design included semantic and phonological competitors
on every trial, so isolating out the semantic and phonological
competition effects required some subtlety. As a result, the lexical
competition effects are only indirectly observed A more direct design
would compare different types of trials: for example, trials with a
target vs. three unrelated images intermixed with trials with a target
vs. a competitor vs. two unrelated images. The trials also used
different kinds of phonological and semantic competitors. For example,
two of the phonological competitors rhymed with the target, so they
could not be included in the analysis of phonological competitors (which
focused on just competitors with the same syllable onset as the target).
The current design limited the number of trials that could be used in
the analyses of the competitors and weakened the power of the analyses.


A related limitation is that the phonological competitors used here are
weak competitors. Adult studies tend to use phonological competitors
with substantial overlap between the target and the competitor. For
example, the landmark study of adults by @Allopenna1998---which showed
that participants' eyetracking probabilities matched lexical activations
from the TRACE model of word recognition [@TRACE]---used
target--cohort pairs that shared a whole syllable: *beaker*--*beetle*, 
*candle*--*candy*, *carrot*--*carriage*, *castle*--*casket*, 
*dollar*--*dolphin*, *paddle*--*padlock*, *pickle*--*picture*, 
*sandal*--*sandwich*. With this degree of overlap, there is much more
phonological and temporal ambiguity for the cohort to build up
activation and compete with the target. In contrast, the words used in
this study were all one syllable and the amount of overlap was limited
to syllable onset (e.g., *flag*--*fly*, *pen*--*pear*). This reduced
overlap limits the degree over temporal ambiguity and thus limits the
degree to which the competitors can participate in lexical competition.
These words were weak phonological competitors, compared to others
studies in the adult literature. As a result, the brief advantage of the
phonological competitor over the unrelated word may *underestimate*
children's sensitivity to phonological competitors: Preschoolers
probably will show much more interference from competitors that have a
larger degree of overlap. Moreover, with more interference from the
competitors, individual differences could emerge more clearly so that
child-level measures like speech perception can predict processing
differences. 

<!-- This limitation, however, should be tempered by the fact that children have smaller vocabularies than adults. The number of familiar words with familiar whole-syllable cohorts is smaller for children, so designing a stimulus-set to tackle these effects is not a trivial matter. -->




A final limitation includes the changes in the experiment procedure over
the course of the longitudinal study. From age 3 to age 4, we
re-recorded the stimuli (with the same original speakers) so that the
noun durations between the two different dialect versions of the
experiment were similar. From age 4 to age 5, we also shortened the
duration of the trials by removing attention-getting prompts (e.g., *this
is fun!*) from the ends of the trials. These small procedural changes
mean that year-to-year differences do not reflect *pure* development
differences. It is implausible, however, that the robust
year-over-changes owe more to procedural changes than a year of learning
and language development.

The findings from this study have implications for our understanding of
word recognition and word learning. The first is the overall
developmental narrative. Preschool children become better at recognizing
words by learning similarities among words and using those similarities
to activate neighborhoods of lexically relevant words. Rather than just
measuring vocabulary size, word recognition reveals how well words have
been integrated into the lexicon. The developmental trends here show
that familiar words become more integrated and more connected over the
preschool years. Even if a child knows a word at age 3 well enough to
recognize or express it, their knowledge of the word will strengthen
over time as the word develops connections to other similar words.

From this perspective, we can think of individual differences in word
recognition as differences in lexical development. Variability in word
recognition diminishes over time, so that differences are more
predictive and discriminating at younger ages. Thus, if we wanted to
intervene on word recognition, these results indicate that early
intervention is better and that intervention should build connections
among words and should target words that build onto existing semantic
and phonological networks. The natural closing of gaps in word
recognition performance with age, however, suggests that word
recognition in and of itself may not be an important intervention
target. Rather, word recognition measures should serve to supplement
other vocabulary measures as an indicator of lexical processing and
lexical integration.
