==> 19-h-check.Rmd <==
Hypothesis check {#aim1-h-check}
========================================================================

\setlength{\parindent}{0in}

Here I revisit my pre-analysis hypotheses.

\bigskip\smallskip

**Children’s accuracy and efficiency of recognizing words will improve
each year.**

Yes. Their curves reached higher heights and showed steeper slopes each
year.

\bigskip\smallskip

**There are stable individual differences in lexical processing of
familiar words such that children who are relatively fast at age 3
remain relatively fast at age 4 and age 5.**

Yes. The rankings of children by lexical processing measures (peak
probability, average probability, linear slope) were concordant over the
three years.

\bigskip\smallskip

**However, the magnitude of these individual differences diminishes
over time, as children converge on a mature level of performance for
this paradigm.**

Yes. I simulated new longitudinal participants based on what the model
learned about the observed children. The range of plausible looking
proportions narrowed each year, so individual differences became less
variable each year.

\bigskip\smallskip

**Consequently, individual differences in word recognition at
age 3, for example, will be more discriminating and predictive of
age-5 language outcomes than differences at age 4 or age 5.**

Yes. Correlations between growth curve features with future vocabulary
measures were strongest for the age-3 growth curve features.

\bigskip\smallskip

**Children will become more sensitive to lexical competitors as they
age, based on the hypothesis that children discover similarities
among words as a consequence of learning more and more words.**

Yes. The advantage of the phonological competitor and semantic
competitor over the unrelated word increased with development.

\bigskip\smallskip

**Children will differ in their sensitivity to lexical competitors,
and these individual differences will correlate with other
child-level measures.**

No evidence to support or refute this hypothesis. I did not find a
relationship between age-3 measures with the phonological or semantic
competitors. In principle, however, one could design a task and derive a
measure from the competitor looking curves that does correlate with
child-level measures.

\setlength{\parindent}{1.5em}
