==> 28-aim2-general-discussion.Rmd <==
General discussion {#aim2-discussion}
=======================================================================

This experiment tested how children responded to familiar words,
one-feature mispronunciations of those familiar words, and unambiguous
novel words. These three types of stimuli allow us to examine features
of children's representations of familiar words and their processing of
unfamiliar words.


A lexical processing account of the results
------------------------------------------------------------------------

Children were sensitive to mispronunciations of familiar words. In terms
of lexical processing, the mispronounced syllable onset leads a child
down a phonological garden path. The "d" in *dirl* activates a
neighborhood a "d"-initial words. The rest of the syllable, however,
provides information needed to activate *girl*. Children therefore are
slower to build up activation of the word because of the onset-mismatch,
and they show less activation overall because the spoken word only
matches the rime of the word. Put differently, they get a late start and
have to work with a poor-fitting form of the word. The finding that
children are better at processing mispronunciations at age 5 suggests
that older children are better able to build up more activation to these
candidate rime words.

Children also processed mispronunciations differently based on the image
they were fixated on. One complication with a simple lexical processing
explanation is that this is a two-image task. Children have ample
time to view each image before the noun onset and build up 
expectations about the words they might hear named. This possibility
leads me to speculate that children might be more uncertain in the
trials where they hear "shoup" while fixated on *soup* because they
build up the activation of *soup*. This prepotent activation would make
the mismatch from the mispronunciation more severe leading to greater
uncertainty. This explanation, however, implies some kind of inhibition
where *soup* suppresses the activation of "sh"-initial words. The
results from Study 1, which do not provide any evidence for changes
in inhibition, make me skeptical of this explanation. Thus, the
effect of children's fixation location on their immediate response to
speech provides an avenue for further research in this area.


A nonword is just a word you haven't learned yet
--------------------------------------------------------------------------

I hypothesized that the nonword condition might be more difficult than
the real word condition, particularly for younger children as seen in
@Bion2013. This prediction did not bear out at all, and indeed, there
was an advantage in the nonword condition in later ages. Part of this
advantage may be a novelty bias. @TRACE_Mispro used the TRACE model of
word recognition [@TRACE] to simulate these kinds of situations. In
one set of simulations, the novel object receives a novelty/salience
boost to resting activation. During presentation of the nonword, none of
the child's known words build up enough activation to overtake the novel
word. In an alternative set of simulations, the novel word is added as
to the lexicon as a low-frequency word, and the absence of competition
of any familiar words causes the novel word to win out. In both of these
accounts, children can quickly associate the novel object with a nonword
because there are no familiar words to interfere with the processing.

The results here probably support both processing accounts. During data
reduction, I separated trials based on initial-fixation location because
children become increasingly likely to start nonword trials on the novel
object. This bias affected 50% of nonword trials at age 3, 55% at age 4,
and 59% at age 5. Thus, there is a novelty preference that for these
trials gets stronger with development. Plus, children are also
learning during this task: At age 5, children were on average were able
to recall the unfamiliar image paired with 5/6 nonwords. This learning
is consistent with the strategy of simulating a nonword as a
low-frequency lexical item.

The findings of @Swingley2007 can help us understand the
mispronunciation retention results. In that study, toddlers were better
able to retain unneighbored nonwords (like *shang* or *meb*) compared to
neighbored nonwords (mispronunciations like *tog* [*dog*] or *gall*
[*ball*]). They concluded that part of fast-referent selection
involves a probability calculation in which children "evaluat[e] the
likelihood that an utterance conveys a new word". In this experiment
(that is, Study 2), we presented an image of the familiar
(mispronounced) object during the mispronunciation trials, which further
reduces the likelihood that the mispronunciation indeed reflects a new,
as-yet unlearned word. Thus, the children in the retention task averaged
around 3--4 mispronunciations correct because those words had a low
likelihood of conveying a new word. In fact, the more children
discounted that probability---that is, the more they looked to the
familiar word on the mispronunciation trials---the less likely they were
to retain the mispronunciation items.

The interference from the familiar word on encoding and retaining the
mispronunciation and image pairing is beneficial for word learning. Suppose
that wordform-object associations are encoded in proportion to the
activation of the wordforms and objects. For the unambiguous nonwords,
the strong activations allow the associations to be built up quickly,
whereas for the mispronunciations, most of activation is spent on the
familiar word, meaning that the association is built up more slowly.
The
use of the word *likelihood* in @Swingley2007 also suggests an analogous
framing using Bayesian terminology: Known words are like priors which
influence (regularize) how unfamiliar words are interpreted, so it takes
more exposure to overcome these priors and encode a mispronunciation as
a new word. In both framings (lexical activation or Bayesian inference),
children showed less retention of the mispronunciations because the
words the children know made the mispronunciations ambiguous and that
ambiguity made each instance of the mispronunciations less informative.
The slowed encoding is a good thing, because spurious associations from
mispronunciations should *not* be learned. @McMurray2012 argues the same
point: "Slow [word] learning may be more optimal in that it prevents
children from committing too strongly to a single (perhaps erroneous)
mapping before they have enough data" [p. 870]. 


Limitations and implications
------------------------------------------------------------------------

The primary limitation for this study is that it applied a procedure
designed for toddlers [@WhiteMorgan2008] on preschoolers. That is to
say, the two-image task was too easy for there to be large
year-over-year developmental changes in children's performance. Children
were successful at recognizing real words and fast-selecting referents
for nonwords at age 3, and by age 4, a quarter of the children performed
at ceiling on the nonword condition. The most difficult condition, based
on the absence of ceiling effects, was the mispronunciation condition in
which children showed much more uncertainty on how to process these
words. For this condition, a developmental trend was also observed where
preschoolers at age 5 had a larger preference for the familiar object on
the trials.

Another limitation here is that the mispronunciations are a particular
kind of mispronunciation: One-feature onset mismatches. That initial
segment sets the stage for the processing of a word as it activates a
word's onset neighbors. Just as it takes longer for a rime to influence
processing of word---as evidence needs to pile up from multiple
compatible sound in order to overcome an initial mismatch---it should
also take longer for a child to recover from an onset-mispronunciation.
Conversely, we might also expect vowel or rime mispronunciation to be
less disruptive for word recognition because of the useful information
on the starting segment. Indeed, @Swingley2009 demonstrates that this is
probably the case. That study tested onset and coda mispronunciations in an
eyetracking task that used two familiar objects. Both adults and
toddlers responded to coda mispronunciations (e.g., "dut" for *duck*) by
showing delayed looks *away* from the (mispronounced) target image. 

As I note in [Appendix \@ref(aim2-mp-items)](#aim2-mp-items), the small
repertoire of mispronunciations is another limitation for this study. It
is conceivable that specific mispronunciations change in severity with
development, even though the canonical form of the word is a familiar
and well known word. In this study, for example, the distance between
the *girl* and *dirl* increased each year, even though *girl* was well
known to three-year-olds, but this distance was driven by children
becoming faster and more efficient at recognizing *girl*. Put another
way, a mispronunciation penalty also reflects children's knowledge of
the canonical word. In this data, the age-5 *rice* receives as many
looks as the age-4 *dirl*. Do children know *rice* less or
accept *dirl* more? With so few items, it is unclear whether these
differences are accidental or systematic.

One implication for this research is that children's recognition of
familiar words and referent selection for nonwords improved a modest
amount over the preschool years. Although these tasks can be done by
toddlers, full mastery does not begin to emerge until age 4. This finding
agrees with the main finding from my analysis of familiar word
recognition: Although children can ostensibly know a word very well,
their recognition and processing of that word can still improve during
preschool.

The finding that mispronunciations were harder to retain than nonwords
also has implications for teaching or intervention. Namely, teaching
words that are confusable with known or relevant words is more difficult
because a child's known words can influence whether the child accepts a
taught word as a novel word.
