==> 29-h-check.Rmd <==
Hypothesis check {#aim2-h-check}
========================================================================

\setlength{\parindent}{0in}

**Children’s accuracy and efficiency of recognizing real words and
fast-associating nonwords will improve each year.**

Yes and no. There were obvious gains for nonwords from 3 to 4 and from 4
to 5 for real words. There are gains, but not "each year". One
complication here is that many children started to hit ceiling
performance by age 4, so there was not much of a developmental gradient
for these conditions.

\bigskip\smallskip

**Performance in real word recognition and fast association of 
nonwords will be highly correlated, based on the hypothesis that the
same process (referent selection) operates in both situations.**

Yes, there were correlated. But not "highly".

\bigskip\smallskip

**Under the alternative hypothesis, real word recognition and fast
referent selection reflect different skills with different
developmental trajectories. Thus, if there is any dissociation
between recognition of real words and nonwords, it will be observed
in younger children.**
    
No---not at all. I observed a dissociation between the two conditions,
but it worked *in the opposite direction* and *at the older ages*. That
is, children at age 4 and age 5 demonstrated an apparent nonword
advantage.

\bigskip\smallskip

**Although these two measures will be correlated, I predict
performance in the nonword condition will be a better predictor of
future vocabulary growth than performance in the real word
condition. This hypothesis is based on the idea that fast referent
selection is a more relevant skill for learning new words than
recognition of known words.**

Yes, peak looking probabilities for the nonword condition at age 3
predicted expressive vocabulary at age 5, and the real word looking
probabilities did not. The size of this effect was rather small,
however.

\bigskip\smallskip

**For the mispronunciations, I predict children with larger
vocabularies (that is, older children) will be more likely to
tolerate a mispronunciation as a production of familiar word
compared to children with smaller vocabularies.**

Yes, older children looked more to the familiar word on average on the
mispronunciation trials.

\bigskip\smallskip

**Mispronunciations that feature later-mastered sounds
(e.g., *rice-wice*) will be more likely to be associated to novel
objects than earlier-mastered sounds (*duck-guck*).**

Not answered. I present the item-level results in in
[Appendix \@ref(aim2-mp-items)](#aim2-mp-items). I chose not to formally
analyze the items because there were only 6 mispronunciations each year,
making it too difficult to generalize about different kinds of
mispronunciations. Plus, children seemed to not know *rice* as well as
the other familiar words which makes the problem of measuring a
mispronunciation penalty more difficult.

\setlength{\parindent}{1.5em}
