==> 95-app-aim2-items.Rmd <==
Items used in Study 2 {#mp-experiment-items}
========================================================================

```{r, include = FALSE}
knitr::read_chunk("./helpers.R")
if (interactive()) source("./helpers.R")
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)
library(readr)
```

```{r helpers, message = FALSE, warning = FALSE}
```


Table \@ref(tab:aim2-items) shows the stimuli design of each year for the
mispronunciation experiment in Study 2. The word list changed between
Age 3 and Age 4 where *dog*/*tog* was replaced with *rice*/*wice*.

```{r aim2-items}
proj_dir <- rprojroot::find_root(rprojroot::is_rstudio_project)
t1 <- "./data-raw/mp_timepoint1_trials.csv.gz"
t2 <- "./data-raw/mp_timepoint2_trials.csv.gz"
t3 <- "./data-raw/mp_timepoint3_trials.csv.gz"

cols <- cols_only(
  Study = col_character(),
  Basename = col_character(),
  Version = col_character(),
  TrialNo = col_integer(),
  WordGroup = col_character(),
  StimType = col_character(),
  TargetWord = col_character(),
  FamiliarImage = col_character(),
  UnfamiliarImage = col_character())
  
df_attributes <- 
  bind_rows(
    readr::read_csv(t1, col_types = cols), 
    readr::read_csv(t2, col_types = cols), 
    readr::read_csv(t3, col_types = cols)) %>% 
  filter(Version == "Standard") %>% 
  mutate(Block = stringr::str_extract(Basename, "Block\\d")) %>% 
  select(-Basename) %>% 
  distinct()

# Long name of experiment condition
condition <- tibble::tribble(
  ~StimType, ~Condition,
  "real", "Correct Production",
  "MP", "Mispronunciation",
  "nonsense", "Nonword"
)

# IPA characters with html entities
ipa <- c(
  "cake" = "kek",
  "gek" = "&#x261;ek",
  "dog" = "d&#x254;&#x261;",
  "tag" = "t&#x254;&#x261;",
  "vef" = "vef",
  "pum" = "p&#x28C;m",
  "soup" = "sup",
  "Sup" = "&#x283;up",
  "duck" = "d&#x28C;k",
  "guk" = "&#x261;&#x28C;k",
  "girl" = "&#x261;&#x025D;l",
  "dRl" = "d&#x025D;l",
  "shoes" = "&#x283;uz",
  "suz" = "suz",
  "San" = "&#x283;æn",
  "Cim" = "&#x2A7;im",
  "giv" = "&#x261;iv",
  "neJ" = "ne&#x2A4;",
  "rice" = "&#x279;a&#x26A;s",
  "waisD" = "wa&#x26A;s",
  "waisC" = "wa&#x26A;s",
  "wais" = "wa&#x26A;s",
  "bep" = "bep"
)

if (knitr:::is_latex_output()) {
  library(kableExtra)
  
  # Convert the HTML entities into symbol calls
  # thanks to 
  # https://tex.stackexchange.com/a/32514/55985
  ipa_names <- names(ipa)
  ipa <- ipa %>% 
    stringr::str_replace_all(
      "&#x(...|....);", 
      "\\\\symbol{\"0\\1}") %>% 
    setNames(ipa_names)
}

english <- c(
  "cake" = "cake",
  "gek" = "gake",
  "dog" = "dog",
  "tag" = "tog",
  "vef" = "vafe",
  "pum" = "pumm",
  "soup" = "soup",
  "Sup" = "shoup",
  "duck" = "duck",
  "guk" = "guck",
  "girl" = "girl",
  "dRl" = "dirl",
  "shoes" = "shoes",
  "suz" = "suze",
  "San" = "shann",
  "Cim" = "cheem",
  "giv" = "geeve",
  "neJ" = "naydge",
  "rice" = "rice",
  "waisD" = "wice",
  "waisC" = "wice",
  "wais" = "wice",
  "bep" = "bape"
)

remove_numbers <- . %>% 
  stringr::str_replace_all("\\d", "")

# Make a table of the stimuli by year, word group, condition
mp_design <- df_attributes %>% 
  distinct(Study, WordGroup, StimType, TargetWord, 
           FamiliarImage, UnfamiliarImage) %>% 
  # Convert from filenames to object words.
  mutate(FamiliarImage = remove_numbers(FamiliarImage), 
         UnfamiliarImage = remove_numbers(UnfamiliarImage),
         TargetWord = TargetWord %>% 
           stringr::str_replace_all("pumA", "pum") %>% 
           stringr::str_replace_all("wais(D|C)", "wais")) %>% 
  distinct() %>% 
  left_join(condition) %>% 
  select(-StimType) %>% 
  arrange(Study, WordGroup, Condition)

# We don't want to repeat rows when the stimuli are the same across different
# years, so make a list of years that featured each stimulus
stim_by_year <- mp_design %>% 
  group_by(WordGroup, Condition) %>% 
  mutate(
    Year = stringr::str_extract(Study, "\\d") %>% as.numeric(),
    Year = Year + 2) %>% 
  summarise(Age = Year %>% unique %>% sort %>% paste0(collapse = ", ")) %>% 
  arrange(Age) %>% 
  ungroup()

mp_design_table <- mp_design %>% 
  left_join(stim_by_year) %>% 
  select(-Study) %>%
  distinct() %>% 
  # left_join(ipa) %>% 
  mutate(
    IPA = ipa[TargetWord],
    `/IPA/` = paste0("\\unicodefont/", IPA, "/\\normalfont"),
    Word = english[TargetWord]) %>% 
  arrange(Age, WordGroup, Condition) %>% 
  select(
    Age, `Word Group` = WordGroup, Condition, 
    `Word`, `/IPA/`, 
    `Familiar` = FamiliarImage, 
    `Unfamiliar` = UnfamiliarImage)

mp_design_table[["Unfamiliar"]] <- 
  mp_design_table[["Unfamiliar"]] %>% 
  stringr::str_replace("bull", "toy creature") %>% 
  stringr::str_replace("holder", "work holder") %>% 
  stringr::str_replace("mixer", "pastry mixer") %>% 
  stringr::str_replace("melon", "horned melon")

mp_design_table[["Condition"]] <- 
  mp_design_table[["Condition"]] %>% 
  stringr::str_replace("Correct Production", "Real word") 

mp_design_table <- mp_design_table %>% 
  group_by(`Word Group`) %>% 
  mutate(
    Age = Age %>% tjmisc::replace_if_same_as_last("")) %>% 
  ungroup() %>% 
  mutate(
    `Word Group` = `Word Group` %>% tjmisc::replace_if_same_as_last("")) 

knitr::kable(
  mp_design_table,
  caption = glue::glue(
    "Items used for the mispronunciation experiment."),
  align = "rclcccc",
  linesep = c("", "", "\\addlinespace"),
  caption.short = "Items used for the mispronunciation experiment.", 
  booktabs = TRUE,
  escape = FALSE)
```
