==> 97-app-related-work.Rmd <==
Related work {#related-work}
========================================================================

In this section, I clarify relationships between this project and other
word recognition research reported from our lab. In short, our lab has
reported results about the two-image and four-image experiments from
cross-sectional samples, describing child-level measures that predict
performance in these tasks. In contrast, my dissertation 1) focuses on
the longitudinal development of word recognition and 2) engages with the
fine-grained details of lexical processing.

@RWLPaper analyzed data from the four-image experiment in 
Study 1. This study featured a diverse cross-sectional sample of 60
children, half of whom received the experiment in African American
English and half received it in Mainstream American English. The sample
ranged in age from 28 to 60 months. The study included data from 23
participants from year 1 of the longitudinal study (i.e., what I refer
to as age 3) in order to enrich parts of the sample demographics. For
this manuscript, we analyzed how vocabulary and maternal education
predicted looking patterns, including relative looks to the semantic and
phonological foils, but with conventional polynomial growth curves. The
use of generalized additive models is an innovation I developed for my
dissertation.

@MPPaper analyzed a different version of the mispronunciation experiment 
on a different sample of children (*n* = 34, 30–46 months old). This
earlier version included both real word and the mispronunciation of the
real word in the same block of trial. For example, a child would hear
“dog” and “tog” during the same session of the experiment. This design
might subtly temper the effect of mispronounced stimuli by allowing the
listener to compare the mispronunciation to its correctly produced
counterpart. The version of the experiment in Study 2 separates
the real words and mispronunciations so that a child never hears a
familiar word and its mispronunciation during the same block of trials.
With this design, there is no explicit point of comparison for the
mispronunciation, and the child has to rely on his or her own lexical
representations when processing these words.

@Mahr2018 was the manuscript I originally authored for
my preliminary examinations. The paper analyzes the same kinds of
relations as @Weisleder2013 which showed that lexical processing
efficiency mediated the effect of language input on future vocabulary
size. Specifically, I asked whether word recognition performance on the
four-image task of Study 1, vocabulary size, and home language
input data from age 3 predicted vocabulary size at
age 4. The paper only examined looks to the familiar image from one
year of the study, so it did not analyze any lexical competition effects
or the development of word recognition within children. In short, we
found that receptive vocabulary was more sensitive to variability in
lexical processing and home language input than expressive vocabulary.
