==> 99-references.Rmd <==
\backmatter

# References {-}

<!-- This manually sets the header for this unnumbered chapter. -->
\markboth{References}{References}

\singlespacing
\sloppy

<!--To remove the indentation of the first entry.-->
\noindent

<!-- To create a hanging indent and spacing between entries. These 
three lines may need to be removed for styles that don't require the 
hanging indent. -->
\setlength{\parindent}{-0.20in}
\setlength{\leftskip}{0.20in}
\setlength{\parskip}{8pt}
